25012 - Functionality moved to separate file.
